renaming basic connector file
