+++ connectors/postgresql.py
@@ -1,4 +1,4 @@
-from .basic_connector import BasicConnector
+from .basic import BasicConnector
 import psycopg2
 import psycopg2.extras
 import psycopg2.pool
